Move types declarations to `nuxt.schema.ts`
